// apps/in-contact-web/src/modules/pso-streaming/enums/timerType.ts
/**
 * @fileoverview TimerType enum
 * @summary Enum for timer types
 * @description Timer types for breaks and emergencies
 */

export enum TimerType {
  LUNCH_BREAK = 'LUNCH_BREAK',
  SHORT_BREAK = 'SHORT_BREAK',
  QUICK_BREAK = 'QUICK_BREAK',
  EMERGENCY = 'EMERGENCY',
  DISCONNECT = 'DISCONNECT',
}